+++ client/src/components/SchemaSubForm.vue
@@ -35,12 +35,11 @@
                                     <InputNumber v-if="field.type == 'number'" :min="(field.property.min > 0)?field.property.min : -Infinity" :max="(field.property.max > 0)?field.property.max : Infinity" v-model="schemainstance.data[index][field.name]" :type="field.type" :placeholder="field.name"></InputNumber>
                                     
                                     <DatePicker v-if="field.type == 'date'" type="date" v-model="schemainstance.data[index][field.name]" :placeholder="(field.property.placeholder !== '') ? field.property.placeholder : field.name"></DatePicker>
-                                    
+                                  
                                     <Select v-if="field.type == 'dropdown'" v-model="schemainstance.data[index][field.name]" :placeholder="(field.property.placeholder !== '') ? field.property.placeholder : field.name">
                                         <Option v-for="dpd in field.property.options" :value="dpd" :key="dpd">{{ dpd }}</Option>
                                     </Select>
                                     <Checkbox v-if="field.type == 'boolean'" v-model="schemainstance.data[index][field.name]">{{field.name}}</Checkbox>
-
                                     <!-- <dynamicinput :type="(field.type) ? field.type : null" :bindmodel="(schemainstance.data[index][field.name]) ? schemainstance.data[index][field.name] : null " :placeholder="(field.property.placeholder !== '') ? field.property.placeholder : field.name" :min="(field.property.min > 0) ? field.property.min : -Infinity" :max="(field.property.max > 0) ? field.property.max : Infinity" :options="(field.property.options) ? field.property.options : null" :field="field"></dynamicinput> -->
                                 </Col>
                             </Row>
@@ -52,26 +51,15 @@
                                 <Col :span="2">
                                     <b>{{field.name}}</b>
                                 </Col>
-                                <!-- <Col :span="20">
-                                <Input v-model="schemainstance.data[index][field.name]" type="text"></Input> -->
-                                <!-- <Input v-model="schemainstance.data[index][field.name]" type="text" :placeholder="(field.property.placeholder !== '') ? field.property.placeholder : field.name"></Input> -->
-                                     <!-- <b>{{field}}</b> -->
-                                     <!-- <a> {{ schemainstance.data[index][field.name] }}</a> -->
-                                <!-- </Col> --> 
                                 <Col :span="21" >
                                     <input class="form-control" type="file" v-if="field.type == 'file'" @change="handleFileChange($event, index, field.name)" :multiple="(field.property.isMultiple)? field.property.isMultiple: false"/>
-                                   <!-- {{fileSize}}  -->
                                     <div v-if="schemainstance.data[index][field.name]" >
-                                      <!-- <p v-for=" val1 in fileSize"> -->
                                       <Progress v-if="stratProgress"  v-bind:percent="fileUploadProgress" :success-percent="30" />
-                                      <!-- {{fileUploadProgress}}  -->
-                                      <!-- </p> --> 
-                                        <div class="" v-for="(val, i) in schemainstance.data[index][field.name]" style="margin-top:10px;">
+                                      <div class="" v-for="(val, i) in schemainstance.data[index][field.name]" style="margin-top:10px;">
                                             <Row>
                                                 <Col :span="23"> <a :href="val" class="list-group-item" target="_blank" style="color:blue;padding:2px 2px;" >{{val}}</a></Col>
                                                 <Col :span="1"><a href="#" style="color:red;float:right"  @click="removeSection(i, schemainstance.data[index][field.name])">&#10005;&nbsp;</a></Col>
                                             </Row>          
-                                            <!-- {{val}} -->
                                         </div>
                                     </div>
                                 </Col>
@@ -94,71 +82,54 @@
   </div>
 </template>
 <script>
-/* eslint-disable */
-import $ from "jquery";
-import SchemaSubForm from "./SchemaSubForm";
-import axios from "axios";
-import moment from "moment";
-// import _ from 'lodash'
+import $ from 'jquery'
+import SchemaSubForm from './SchemaSubForm'
+import axios from 'axios'
+import moment from 'moment'
+import _ from 'lodash'
 // let jumperLinks = [];
-var AWS = require("aws-sdk");
+var AWS = require('aws-sdk')
 AWS.config.update({
   accessKeyId: process.env.accesskey,
   secretAccessKey: process.env.secretkey
-});
-AWS.config.region = "us-west-2";
-
+})
+AWS.config.region = 'us-west-2'
 export default {
-  name: "schemasubform",
-  props: ["schemainstance"],
-  data() {
+  name: 'schemasubform',
+  props: ['schemainstance'],
+  data () {
     return {
-      fileNumber:0,
+      fileNumber: 0,
       fileSize: 0,
       fileUploadProgress: 0,
       stratProgress: false,
       jumperLinks: []
-    };
+    }
   },
   components: {
     schemasubform: SchemaSubForm
   },
   methods: {
-    async handleFileChange(e, index, fieldName) {
-      // return new Promise((resolve, reject) => {
-      // console.log("fuunction caliing");
-      let self = this;
-      var files = e.target.files || e.dataTransfer.files;
-      let allFiles = [];                                                                                                                                                      
-      // let allFiles
+    async handleFileChange (e, index, fieldName) {
+      let self = this
+      var files = e.target.files || e.dataTransfer.files
+      let allFiles = []
       if (files.length > 0) {
-        // console.log("files", files[0]);                                                                                                                                                                     
         for (let i = 0; i < files.length; i++) {
-          self.stratProgress = true;
-          // console.log(files.name)                                                                                                                            );
-          // console.log("for loop");
+          self.stratProgress = true
           self.fileSize = files.length
-          // console.log(fileSize);
-          // console.log("file");
-          // console.log(files[i]);
-          // console.log("for loop");
-          let abc = await this.uploadToAWS(files[i], i);
-          allFiles.push(abc);
+          let abc = await this.uploadToAWS(files[i], i)
+          allFiles.push(abc)
         }
       }
-      // self.schemainstance.data[index][fieldName] = allFiles
-      self.schemainstance.data[index][fieldName] = _.concat(self.schemainstance.data[index][fieldName],allFiles);
-      // console.log(allFiles);
-      // console.log(self.schemainstance.data[index][fieldName]);
+      self.schemainstance.data[index][fieldName] = _.concat(self.schemainstance.data[index][fieldName], allFiles)
     },
-    uploadToAWS(file, i) {
-      let self = this;
-      
-      // console.log(fileSize);
+    uploadToAWS (file, i) {
+      let self = this
       return new Promise((resolve, reject) => {
         let bucket = new AWS.S3({
-          params: { Bucket: "airflowbucket1/obexpense/expenses" }
-        });
+          params: { Bucket: 'airflowbucket1/obexpense/expenses' }
+        })
         var params = {
           Key:
             moment()
@@ -168,104 +139,95 @@ export default {
             file.name,
           ContentType: file.type,
           Body: file
-        };
+        }
         bucket
           .upload(params)
-          .on("httpUploadProgress", function(evt) {
-            // console.log(i);
-            self.fileUploadProgress = parseInt(evt.loaded * 100 / evt.total);
+          .on('httpUploadProgress', function (evt) {
+            self.fileUploadProgress = parseInt(evt.loaded * 100 / evt.total)
           })
-          .send(function(err, data) {
+          .send(function (err, data) {
             if (err) {
-              alert(err);
-              reject(err);
+              alert(err)
+              reject(err)
             } else {
-              // console.log("done");
-              // await allFiles.push(data.Location)
-              // console.log(data.Location);
-              self.fileNumber = i+1;
-              self.fileUploadProgress = 0;
-              self.stratProgress = false;
-              resolve(data.Location);
-              // console.log(allFiles.push(data.Location))
+              self.fileNumber = i + 1
+              self.fileUploadProgress = 0
+              self.stratProgress = false
+              resolve(data.Location)
             }
-          });
-        //  }
-      });
-      // })
+          })
+      })
     },
-    removeSection: function(index, imgarray) {
-    	console.log(index);
-			imgarray.splice(index, 1)
+    removeSection: function (index, imgarray) {
+      imgarray.splice(index, 1)
     },
-    getValidationProps(index, fieldName) {
-      return "data[" + index + "][" + fieldName + "]";
+    getValidationProps (index, fieldName) {
+      return 'data[' + index + '][' + fieldName + ']'
     },
-    async getChildData(id) {
-      // alert(id)
-      var arrObj = [];
-      var self = this;
+    async getChildData (id) {
+      var arrObj = []
+      var self = this
       await axios
-        .get("https://api.flowzcluster.tk/eng/schema/" + id)
+        .get('https://api.flowzcluster.tk/eng/schema/' + id)
         .then(async response => {
-          var _res = response.data;
-          var obj = {};
-          // obj.id = self.getGuid();  // for guid for perticular row
+          var _res = response.data
+          var obj = {}
+          // obj.id = self.getGuid()  // for guid for perticular row
           // obj.database = _res.database
           // obj.Schemaid = _res._id
           for (let v of _res.entity) {
             if (v.customtype) {
-              obj[v.name] = await self.getChildData(v.type);
+              obj[v.name] = await self.getChildData(v.type)
             } else {
-              if (v.type === "number") {
-                if (v.property.defaultValue !== "") {
-                  obj[v.name] = v.property.defaultValue;
+              if (v.type === 'number') {
+                if (v.property.defaultValue !== '') {
+                  obj[v.name] = v.property.defaultValue
                 } else {
-                  if (v.property.min !== 0 && v.property.min !== "") {
-                    obj[v.name] = v.property.min;
+                  if (v.property.min !== 0 && v.property.min !== '') {
+                    obj[v.name] = v.property.min
                   } else {
-                    obj[v.name] = 1;
+                    obj[v.name] = 1
                   }
                 }
-              } else if (v.type === "boolean") {
+              } else if (v.type === 'boolean') {
                 if (
-                  v.property.defaultValue !== "" ||
-                  v.property.defaultValue === "true"
+                  v.property.defaultValue !== '' ||
+                  v.property.defaultValue === 'true'
                 ) {
-                  obj[v.name] = true;
+                  obj[v.name] = true
                 } else {
-                  obj[v.name] = false;
+                  obj[v.name] = false
                 }
-              } else if (v.type === "file") {
-                obj[v.name] = [];
+              } else if (v.type === 'file') {
+                obj[v.name] = []
               } else {
-                if (v.property.defaultValue !== "") {
-                  obj[v.name] = v.property.defaultValue;
+                if (v.property.defaultValue !== '') {
+                  obj[v.name] = v.property.defaultValue
                 } else {
-                  obj[v.name] = "";
+                  obj[v.name] = ''
                 }
               }
             }
           }
-          arrObj.push(obj);
+          arrObj.push(obj)
         })
         .catch(error => {
-          console.log("Error: ", error);
-        });
-      return arrObj;
+          console.log('Error: ', error)
+        })
+      return arrObj
     },
-    getObject(eIndex, dataIndex, fname, ftype) {
-      var obj = {};
-      obj.data = this.schemainstance.data[dataIndex][fname];
-      obj.entity = this.schemainstance.entity[eIndex].entity[0].entity;
-      let indexx = $.inArray(fname, this.jumperLinks);
+    getObject (eIndex, dataIndex, fname, ftype) {
+      var obj = {}
+      obj.data = this.schemainstance.data[dataIndex][fname]
+      obj.entity = this.schemainstance.entity[eIndex].entity[0].entity
+      let indexx = $.inArray(fname, this.jumperLinks)
       if (indexx === -1) {
-        this.jumperLinks.push(fname);
+        this.jumperLinks.push(fname)
       }
-      return obj;
+      return obj
     },
-    async handleAdd(eIndex, dataIndex, ent, data, fname) {
-      var self = this;
+    async handleAdd (eIndex, dataIndex, ent, data, fname) {
+      var self = this
       // var self.$refs['formSchema'][0].validate()
       // self.$refs['formSchema'][0].validate((valid) => {
       //   alert(1)
@@ -275,134 +237,132 @@ export default {
       //     this.$Message.error('Error!')
       //   }
       // })
-      var obj = {};
-      // obj.id = this.getGuid();
+      var obj = {}
+      // obj.id = this.getGuid()
       // alert(ent.database)
       // obj.database = ent.database
       // obj.Schemaid = ent._id
       for (let v of ent.entity) {
         if (v.customtype) {
-          obj[v.name] = await self.getChildData(v.type);
+          obj[v.name] = await self.getChildData(v.type)
         } else {
-          if (v.type === "number") {
-            if (v.property.defaultValue !== "") {
-              obj[v.name] = v.property.defaultValue;
+          if (v.type === 'number') {
+            if (v.property.defaultValue !== '') {
+              obj[v.name] = v.property.defaultValue
             } else {
-              if (v.property.min !== 0 && v.property.min !== "") {
-                obj[v.name] = v.property.min;
+              if (v.property.min !== 0 && v.property.min !== '') {
+                obj[v.name] = v.property.min
               } else {
-                obj[v.name] = 1;
+                obj[v.name] = 1
               }
             }
-          } else if (v.type === "boolean") {
+          } else if (v.type === 'boolean') {
             if (
-              v.property.defaultValue !== "" ||
-              v.property.defaultValue === "true"
+              v.property.defaultValue !== '' ||
+              v.property.defaultValue === 'true'
             ) {
-              obj[v.name] = true;
+              obj[v.name] = true
             } else {
-              obj[v.name] = false;
+              obj[v.name] = false
             }
-          } else if (v.type === "file") {
-            obj[v.name] = [];
+          } else if (v.type === 'file') {
+            obj[v.name] = []
           } else {
-            if (v.property.defaultValue !== "") {
-              obj[v.name] = v.property.defaultValue;
+            if (v.property.defaultValue !== '') {
+              obj[v.name] = v.property.defaultValue
             } else {
-              obj[v.name] = "";
+              obj[v.name] = ''
             }
           }
         }
       }
-      this.schemainstance.data[dataIndex][fname].push(obj);
-      // console.log('schemainstance: ', this.schemainstance.data)
+      this.schemainstance.data[dataIndex][fname].push(obj)
     },
     // getGuid () {
-    //   return (this.S4() + this.S4() + "-" + this.S4() + "-4" + this.S4().substr(0,3) + "-" + this.S4() + "-" + this.S4() + this.S4() + this.S4()).toLowerCase()
+    //   return (this.S4() + this.S4() + '-' + this.S4() + '-4' + this.S4().substr(0,3) + '-' + this.S4() + '-' + this.S4() + this.S4() + this.S4()).toLowerCase()
     // },
     // S4() {
-    //     return (((1+Math.random())*0x10000)|0).toString(16).substring(1);
+    //     return (((1+Math.random())*0x10000)|0).toString(16).substring(1)
     // },
-    getObjectType(type) {
+    getObjectType (type) {
       return (
         [
-          "text",
-          "email",
-          "number",
-          "phone",
-          "boolean",
-          "date",
-          "dropdown",
-          "file"
+          'text ',
+          'email ',
+          'number ',
+          'phone ',
+          'boolean ',
+          'date ',
+          'dropdown ',
+          'file '
         ].indexOf(type) === -1
-      );
+      )
     },
-    createRules(row) {
-      let rules = [];
-      if (row.type === "email") {
-        rules.push({ type: "email", message: "This field is email type." });
+    createRules (row) {
+      let rules = []
+      if (row.type === 'email') {
+        rules.push({ type: 'email', message: 'This field is email type.' })
       }
-      if (row.type === "phone") {
+      if (row.type === 'phone') {
         rules.push({
-          type: "number",
+          type: 'number',
           validator: (rule, value, callback) => {
             if (!value) {
-              return callback(new Error("Please input the value"));
+              return callback(new Error('Please input the value'))
             }
             if (
               value.match(/^\(?([0-9]{3})\)?[-. ]?([0-9]{3})[-. ]?([0-9]{4})$/)
             ) {
-              callback();
+              callback()
             } else {
-              callback(new Error("Please input phone no"));
+              callback(new Error('Please input phone no'))
             }
           }
-        });
+        })
       }
       if (row.property.optional === false) {
         rules.push({
           required: true,
-          message: "This field is required.",
-          trigger: "blur"
-        });
+          message: 'This field is required.',
+          trigger: 'blur'
+        })
       }
-      if (row.property.min > 0 && row.type === "text") {
+      if (row.property.min > 0 && row.type === 'text') {
         rules.push({
-          type: "string",
+          type: 'string',
           min: row.property.min,
-          message: "min" + row.property.min + " length req"
-        });
+          message: 'min' + row.property.min + ' length req'
+        })
       }
-      if (row.property.max > 0 && row.type === "text") {
+      if (row.property.max > 0 && row.type === 'text') {
         rules.push({
-          type: "string",
+          type: 'string',
           max: row.property.max,
-          message: "max length " + row.property.max + " required."
-        });
+          message: 'max length ' + row.property.max + 'required.'
+        })
       }
       if (row.property.allowedValue.length > 0) {
-        rules.push({ type: "enum", enum: row.property.allowedValue });
+        rules.push({ type: 'enum', enum: row.property.allowedValue })
       }
-      return rules;
+      return rules
     },
-    handleEdit(row) {},
-    handleRemove(index) {
+    handleEdit (row) {},
+    handleRemove (index) {
       this.$Modal.confirm({
-        title: "Confirm",
-        content: "<p>Are you sure you want to delete?</p>",
+        title: 'Confirm',
+        content: '<p>Are you sure you want to delete?</p>',
         onOk: () => {
-          this.schemainstance.data.splice(index, 1);
+          this.schemainstance.data.splice(index, 1)
         },
         onCancel: () => {}
-      });
+      })
     }
   },
-  mounted() {},
-  created() {}
-};
+  mounted () {},
+  created () {}
+}
 </script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
+<!-- Add 'scoped' attribute to limit CSS to this component only -->
 <style scoped>
 .ui-card {
   background-color: #fff;
